requirements.txt
canvasapi==3.2.0
canvaslms==4.0
Django==4.2.13
requests==2.31.0
python-dateutil~=2.9.0.post0